fixup docs, move timestamp to _internal.timestamp
